backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
python-multipart
pydantic